requirements.txt
Flask==3.1.0
python-barcode==0.15.1
Pillow==11.0.0
